requirements.txt
pandas>=2.2.0
pyarrow>=15.0.0
numpy>=1.26.0
tqdm>=4.66.0
nltk>=3.8.1
scikit-learn>=1.3.0
rank-bm25>=0.2.2
faiss-cpu>=1.7.4
sentence-transformers>=3.0.0
transformers>=4.41.0
rapidfuzz>=3.6.1
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
